+++ lib/lamby/templates/rest/app.rb
@@ -1,0 +1,10 @@
+require_relative 'config/boot'
+require 'lamby'
+require_relative 'config/application'
+require_relative 'config/environment'
+
+$app = Rack::Builder.new { run Rails.application }.to_app
+
+def handler(event:, context:)
+  Lamby.handler $app, event, context, rack: :rest
+end